pyproject.toml
[tool.poetry]
name = "checkout-event-data-pipeline"
version = "0.1.0"
description = ""
authors = ["kaushik911 <kaushikshamantha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
faker = "^26.0.0"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
